fix(eap) Separate the RPC code from the other snuba code
Move the RPC endpoint and code into a separate file. Also create a new threadpool for the RPC
to use. This won't result in more work, queries are sent either to the RPC or the SnQL endpoint
but not both.
